DateRangePill: seed defaults when activated from the per-topic adder
When the operator activates the date range from the `+` inside
the Time chip (rather than the top-level Filters adder), the pill
got an empty `{}` instead of the URL-seeded defaults from
`seedDateRangeFromUrl`. Both adders now hit the same helper —
extracted to `Filters/seedDateRange.ts` so the seeding rule (year
alone → Jan 1–Dec 31; year+month → first–last of month; full
date → that day on both sides) stays single-sourced.
