+++ react-app/src/components/Gallery/Filters/index.tsx
@@ -13,40 +13,7 @@ import Topic from "./Topic";
 import filter, { type Filters as FiltersT } from "../../../lib/filter";
 import type { UniqueValues } from "../../../lib/stats";
 import { useBetaStore, useFiltersStore } from "../../../stores";
-import { type DateRange } from "../../../stores/filters";
-
-// Pre-fill the date-range pill based on the URL's year / month /
-// day when the operator activates it from the Time-topic adder.
-// Browser native `<input type="date">` doesn't expose a "default
-// month" attribute, so the only way to anchor the calendar to the
-// currently-viewed slice is to seed the inputs themselves. The
-// operator can clear / adjust from there.
-const seedDateRangeFromUrl = (
-  year?: string,
-  month?: string,
-  day?: string
-): DateRange => {
-  if (year && month && day) {
-    const ymd = `${year}-${month.padStart(2, "0")}-${day.padStart(2, "0")}`;
-    return { from: ymd, to: ymd };
-  }
-  if (year && month) {
-    const m = month.padStart(2, "0");
-    const lastDay = new Date(
-      Number(year),
-      Number(month),
-      0
-    ).getDate();
-    return {
-      from: `${year}-${m}-01`,
-      to: `${year}-${m}-${String(lastDay).padStart(2, "0")}`,
-    };
-  }
-  if (year) {
-    return { from: `${year}-01-01`, to: `${year}-12-31` };
-  }
-  return {};
-};
+import seedDateRangeFromUrl from "./seedDateRange";
 
 interface CountryData {
   getName(code: string, lang: string): string | undefined;
